Create SmartConfig_PushButton.ino
This is an Arduino sketch to use a push button to run the TI smart config app to change your wifi password. I'm using the CC3000 library from adafruit.com. The program will first attempt to connect to the existing WiFi settings. There are two ways to run the smart config app. First if the arduino has trouble connecting to an existing network, then the smart config will run and if you press the button to set a flag for the main loop to run the smart config.